revert common jwt, specs
