+++ spec/lib/bpds/jwt_generator_spec.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require 'bpds/jwt_encoder'
+require 'bpds/jwt_generator'
 
-RSpec.describe BPDS::JwtEncoder do
-  describe '#get_token' do
+RSpec.describe BPDS::JwtGenerator do
+  describe '#encode_jwt' do
     it 'returns a token with required fields' do
-      encoded_jwt = BPDS::JwtEncoder.new.get_token
+      encoded_jwt = BPDS::JwtGenerator.new.encode_jwt
       decoded_jwt = JWT.decode(encoded_jwt, Settings.bpds.jwt_secret, true, {
                                  typ: 'JWT',
                                  alg: 'HS256'
